pyproject.toml
[tool.poetry]
name = "safebot-chainlit"
version = "0.1.0"
description = "SafeBot - Sistema Inteligente de Segurança do Trabalho com Chainlit"
authors = ["SafeBot Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
chainlit = "^2.0.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
chromadb = "^0.5.0"
openai = "^1.102.0"
python-dotenv = "^1.0.0"
pypdf = "^5.1.0"
tiktoken = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
